% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo into an image}
\usage{
add_logo(plot_path, logo_path, logo_position = c("top right", "top left",
  "bottom right", "bottom left"), logo_scale = 10)
}
\arguments{
\item{plot_path}{image path.}

\item{logo_path}{logo image path.}

\item{logo_position}{the position of logo within the image.}

\item{logo_scale}{the size of logo relative to the image.}
}
\value{
Image.
}
\description{
Add a logo into an image
}
\examples{
\donttest{add_logo(
    plot_path = system.file("extdata", "logo.png", package = "add2ggplot"),
    logo_path = system.file("extdata", "jiaxiang.png", package = "add2ggplot"),
    logo_position = "bottom right",
    logo_scale = 5
)}
}
