% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey_and_red.R
\docType{data}
\name{grey_one}
\alias{grey_one}
\title{The chosen color in grey}
\format{An object of class \code{character} of length 1.}
\usage{
grey_one
}
\description{
The chosen color in grey
}
\keyword{datasets}
