% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_colors.R
\name{more_colors}
\alias{more_colors}
\title{Use a palette with 24 colors}
\usage{
more_colors(n = 24)
}
\arguments{
\item{n}{the max number of color you choose}
}
\value{
The vector of the colors you choose
}
\description{
\code{more_colors} helps users to build a more robust palette serving ggplot objects.
}
\author{
Jiaxiang Li
}
