% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey_and_red.R
\docType{data}
\name{red_two}
\alias{red_two}
\title{The chosen color in red}
\format{An object of class \code{character} of length 1.}
\usage{
red_two
}
\description{
The chosen color in red
}
\keyword{datasets}
