% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic2.R
\name{theme_classic2}
\alias{theme_classic2}
\title{Another classic theme}
\usage{
theme_classic2()
}
\value{
ggplot object.
}
\description{
# Inspired by
https://www.datacamp.com//courses/intermediate-data-visualization-with-ggplot2
}
\examples{
datasets::mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(mpg, disp)) +
  ggplot2::geom_point() +
  theme_classic2()
}
