% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey_and_red.R
\name{theme_grey_and_red}
\alias{theme_grey_and_red}
\title{Use the ggplot theme in grey and red}
\usage{
theme_grey_and_red()
}
\value{
The function to use \code{+} in the ggplot workflow.
}
\description{
Use the ggplot theme in grey and red
}
\author{
Jiaxiang Li
}
