% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_white.R
\name{theme_white}
\alias{theme_white}
\title{A ggplot theme with a white background.}
\usage{
theme_white()
}
\value{
The function to use \code{+} in the ggplot workflow.
}
\description{
A ggplot theme with a white background.
}
\examples{
datasets::mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(mpg, disp)) +
  ggplot2::geom_point() +
  theme_white()
}
