% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey_and_red.R
\docType{data}
\name{white_one}
\alias{white_one}
\title{The chosen color in white}
\format{An object of class \code{character} of length 1.}
\usage{
white_one
}
\description{
The chosen color in white
}
\keyword{datasets}
