% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo into an image}
\usage{
add_logo(plot_path, logo_path, logo_position = c("top right", "top left",
  "bottom right", "bottom left"), logo_scale = 10)
}
\arguments{
\item{plot_path}{image path.}

\item{logo_path}{logo image path.}

\item{logo_position}{the position of logo within the image.}

\item{logo_scale}{the size of logo relative to the image.}
}
\description{
Add a logo into an image
}
\examples{
\dontrun{add_logo(}
\dontrun{    plot_path = "man/figures/logo.png",}
\dontrun{    logo_path = "man/figures/jiaxiang.png",}
\dontrun{    logo_position = "bottom right",}
\dontrun{    logo_scale = 5)}
}
