% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic2.R
\name{theme_classic2}
\alias{theme_classic2}
\title{Another classic theme}
\usage{
theme_classic2()
}
\description{
# Inspired by
https://www.datacamp.com//courses/intermediate-data-visualization-with-ggplot2
}
