% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_du_bois.R
\name{theme_du_bois}
\alias{theme_du_bois}
\title{Paste Kindle Notes in an elegant way}
\usage{
theme_du_bois(font_name = "")
}
\value{
function
}
\description{
Paste Kindle Notes in an elegant way
}
\author{
Jiaxiang Li
}
