% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey_and_red.R
\name{theme_grey_and_red}
\alias{theme_grey_and_red}
\title{Use the ggplot theme in grey and red}
\usage{
theme_grey_and_red()
}
\value{
The function to use \code{+} in the ggplot workflow.
}
\description{
Use the ggplot theme in grey and red
}
\examples{
datasets::mtcars \%>\%
  ggplot2::ggplot(ggplot2::aes(mpg, disp)) +
  ggplot2::geom_point() +
  theme_grey_and_red()
}
\author{
Jiaxiang Li
}
